==> EDA/2_Customer_Behavior_Profile_Analysis.ipynb <==
This code aggregates the overall customer behavior profile for each month created from the notebook: Customer_Behavior_Profile_Month.ipynb

# Loading basic needed libraries
import pandas as pd
import numpy as np
import gc
from functools import reduce
import datetime as dt
from datetime import date

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3')

import plotly.express as px

#### Creating full customer profile

# Reading customer overall profiles from S3
df_oct = pd.read_csv('s3://myaws-capstone-bucket/data/2019-Oct/customer_profile.csv')
df_nov = pd.read_csv('s3://myaws-capstone-bucket/data/2019-Nov/customer_profile.csv')
df_dec = pd.read_csv('s3://myaws-capstone-bucket/data/2019-Dec/customer_profile.csv')
df_jan = pd.read_csv('s3://myaws-capstone-bucket/data/2020-Jan/customer_profile.csv')
df_feb = pd.read_csv('s3://myaws-capstone-bucket/data/2020-Feb/customer_profile.csv')
df_mar = pd.read_csv('s3://myaws-capstone-bucket/data/2020-Mar/customer_profile.csv')
df_apr = pd.read_csv('s3://myaws-capstone-bucket/data/2020-Apr/customer_profile.csv')

# Concatanating dfs together to create one full customer profile for the 7 months
data_frames = [df_oct, df_nov, df_dec, df_jan, df_feb, df_mar, df_apr]
cust_profile = pd.concat(data_frames)
cust_profile.nunique()

# Rank column based on month customer shopped
cust_profile.loc[(cust_profile['month'] == '2019-Oct'),'month_rank'] = 7
cust_profile.loc[(cust_profile['month'] == '2019-Nov'),'month_rank'] = 6
cust_profile.loc[(cust_profile['month'] == '2019-Dec'),'month_rank'] = 5
cust_profile.loc[(cust_profile['month'] == '2020-Jan'),'month_rank'] = 4
cust_profile.loc[(cust_profile['month'] == '2020-Feb'),'month_rank'] = 3
cust_profile.loc[(cust_profile['month'] == '2020-Mar'),'month_rank'] = 2
cust_profile.loc[(cust_profile['month'] == '2020-Apr'),'month_rank'] = 1

cust_profile.nunique()

# Saving results with month data added
cust_profile.to_csv('s3://myaws-capstone-bucket/data/cust_profile_full_month_agg.csv',index=False)

# Transforming date columns to datetime type
cust_profile["max_activity_date"] = cust_profile["max_activity_date"].astype('datetime64[ns]')
cust_profile["min_activity_date"] = cust_profile["min_activity_date"].astype('datetime64[ns]')

cust_profile["min_view_date"] = cust_profile["min_view_date"].astype('datetime64[ns]')
cust_profile["max_view_date"] = cust_profile["max_view_date"].astype('datetime64[ns]')

cust_profile["min_cart_date"] = cust_profile["min_cart_date"].astype('datetime64[ns]')
cust_profile["max_cart_date"] = cust_profile["max_cart_date"].astype('datetime64[ns]')

cust_profile["max_purchase_date"] = cust_profile["max_purchase_date"].astype('datetime64[ns]')
cust_profile["min_purchase_date"] = cust_profile["min_purchase_date"].astype('datetime64[ns]')

# Grabbing max dates based on month rank
max_activity_dates = cust_profile.sort_values('max_activity_date',ascending = False).groupby('user_id').head(1)
max_activity_dates = max_activity_dates[['user_id','max_activity_date']]

max_view_dates = cust_profile.sort_values('max_view_date',ascending = False).groupby('user_id').head(1)
max_view_dates = max_view_dates[['user_id','max_view_date']]

max_cart_dates = cust_profile.sort_values('max_cart_date',ascending = False).groupby('user_id').head(1)
max_cart_dates = max_cart_dates[['user_id','max_cart_date']]

max_purchase_dates = cust_profile.sort_values('max_purchase_date',ascending = False).groupby('user_id').head(1)
max_purchase_dates = max_purchase_dates[['user_id','max_purchase_date']]

# merging_max_dates
data_frames = [max_activity_dates,max_view_dates, max_cart_dates, max_purchase_dates]
max_dates = reduce(lambda  left,right: pd.merge(left,right,on=['user_id'],how='outer'), data_frames)

# Grabbing min dates based on month rank
min_activity_dates = cust_profile.sort_values('min_activity_date',ascending = True).groupby('user_id').head(1)
min_activity_dates = min_activity_dates[['user_id','min_activity_date']]

min_view_dates = cust_profile.sort_values('min_view_date',ascending = True).groupby('user_id').head(1)
min_view_dates = min_view_dates[['user_id','min_view_date']]

min_cart_dates = cust_profile.sort_values('min_cart_date',ascending = True).groupby('user_id').head(1)
min_cart_dates = min_cart_dates[['user_id','min_cart_date']]

min_purchase_dates = cust_profile.sort_values('min_purchase_date',ascending = True).groupby('user_id').head(1)
min_purchase_dates = min_purchase_dates[['user_id','min_purchase_date']]

# merging_min_dates
data_frames = [min_activity_dates,min_view_dates, min_cart_dates, min_purchase_dates]
min_dates = reduce(lambda  left,right: pd.merge(left,right,on=['user_id'],how='outer'), data_frames)

# Obtaining max possible date from data 
column = df_apr["max_activity_date"].astype('datetime64[ns]')
max_date = column.max() 
# Transforming date columns to datetime format
max_dates[['max_activity_date','max_view_date','max_cart_date', 'max_purchase_date']] = max_dates[['max_activity_date','max_view_date','max_cart_date', 'max_purchase_date']].astype('datetime64[ns]')
min_dates[['min_activity_date','min_view_date','min_cart_date', 'min_purchase_date']] = min_dates[['min_activity_date','min_view_date','min_cart_date', 'min_purchase_date']].astype('datetime64[ns]')

customer_dates = pd.merge(max_dates, min_dates, on=["user_id"], how='inner')

customer_dates.head()

# Calculating days since each event and customer activity
customer_dates['cust_retailer_age'] = (max_date - customer_dates['min_activity_date']).dt.days
customer_dates['days_since_last_activity'] = (max_date - customer_dates['max_activity_date']).dt.days

customer_dates['first_view_age'] = (max_date - customer_dates['min_view_date']).dt.days
customer_dates['days_since_last_view'] = (max_date - customer_dates['max_view_date']).dt.days

customer_dates['first_cart_age'] = (max_date - customer_dates['min_cart_date']).dt.days
customer_dates['days_since_last_cart'] = (max_date - customer_dates['max_cart_date']).dt.days


customer_dates['first_purchase_age'] = (max_date - customer_dates['min_purchase_date']).dt.days
customer_dates['days_since_last_purchase'] = (max_date - customer_dates['max_purchase_date']).dt.days

customer_dates[['cust_retailer_age',
       'days_since_last_activity', 'first_view_age', 'days_since_last_view',
       'first_cart_age', 'days_since_last_cart', 'first_purchase_age',
       'days_since_last_purchase']] = customer_dates[['cust_retailer_age',
       'days_since_last_activity', 'first_view_age', 'days_since_last_view',
       'first_cart_age', 'days_since_last_cart', 'first_purchase_age',
       'days_since_last_purchase']].fillna(value=9999)# Replacing nan with 9999 to symbolize they have not done said action

customer_dates.head()

# Grabbing min and max spent
max_spent = cust_profile.groupby(['user_id']).max_spent.agg(max_spent=np.max).reset_index()
min_spent = cust_profile.groupby(['user_id']).min_spent.agg(min_spent=np.min).reset_index()

customer_spent = pd.merge(min_spent, max_spent, on=["user_id"], how='inner')
customer_spent.head()

# Merging spent and date dfs
customer_agg = pd.merge(customer_spent, customer_dates[['user_id','cust_retailer_age',
       'days_since_last_activity', 'first_view_age', 'days_since_last_view',
       'first_cart_age', 'days_since_last_cart', 'first_purchase_age',
       'days_since_last_purchase']], on=["user_id"], how='inner')
customer_agg.nunique()

# Aggregating numeric columns from cust_profile
customer_views = pd.DataFrame(cust_profile.groupby(['user_id'])['view_count'].agg('sum')).reset_index()
customer_views.columns = ['user_id','total_view']# Renaming columns
customer_carts = pd.DataFrame(cust_profile.groupby(['user_id'])['cart_count'].agg('sum')).reset_index()
customer_carts.columns = ['user_id','total_cart_add']# Renaming columns
customer_purchases = pd.DataFrame(cust_profile.groupby(['user_id'])['purchase_count'].agg('sum')).reset_index()
customer_purchases.columns = ['user_id','total_purchases']# Renaming columns
customer_sessions = pd.DataFrame(cust_profile.groupby(['user_id'])['session_count'].agg('sum')).reset_index()
customer_sessions.columns = ['user_id','total_sessions']# Renaming columns
customer_spent = pd.DataFrame(cust_profile.groupby(['user_id'])['spent'].agg('sum')).reset_index()
customer_spent.columns = ['user_id','total_spent']# Renaming columns

# Merging event counts dfs together
data_frames = [customer_views, customer_carts, customer_purchases, customer_sessions, customer_spent]
customer_profile = reduce(lambda  left,right: pd.merge(left,right,on=['user_id'],how='outer'), data_frames)
customer_profile.nunique()

customer_profile = pd.merge(customer_profile, customer_agg, on=["user_id"], how='inner')
customer_profile.head()

customer_profile.nunique()

customer_profile.shape

# Saving results in S3
customer_profile.to_csv('s3://myaws-capstone-bucket/data/customer_profile_full.csv',index=False)

#### Analyzing customer behavior

pd.options.display.float_format = '{:.2f}'.format
customer_profile.describe()

none_purchase_custs = customer_profile.loc[(customer_profile['total_spent'] == 0)]
none_purchase_custs.nunique()

none_purchase_custs.describe()

11881941/13699114

purchase_custs = customer_profile.loc[(customer_profile['total_spent'] > 0)]
purchase_custs.nunique()

purchase_custs.describe()

1817173/13699114*100

# Saving results into S3
purchase_custs.to_csv('s3://myaws-capstone-bucket/data/customers_of_focus.csv',index=False)
