==> EDA/4_purchase_dataset_gathering.ipynb <==
This notebook loads the 7 months (from October 2019 to April 2020) of data only focusing on the purchase events for each user, taking their event_time, category_code and brand. At the end we will have one full dataframe containing the purchases of each user with their date. 

The resulting dataframe will be used for Basket Market Analysis along with other product recommendation models based on sequences of purchase such as RNN and CNN recommendation models and Neural Collaborative Filtering model.

# Loading basic needed libraries
import pandas as pd
import gc

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

#### Loading and preparing data

# Reading csv file from S3 - Selecting specific columns
df_oct = pd.read_csv('s3://predictive-maintenance-bucket/data/2019-Oct.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code','category_id','brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_oct = df_oct.dropna(subset=['category_code', 'brand'])
df_oct = df_oct.loc[df_oct['event_type'] == 'purchase']
df_oct.nunique()

# Reading csv file from S3 - Selecting specific columns
df_nov = pd.read_csv('s3://predictive-maintenance-bucket/data/2019-Nov.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code','category_id', 'brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_nov = df_nov.dropna(subset=['category_code', 'brand'])
df_nov = df_nov.loc[df_nov['event_type'] == 'purchase']
df_nov.nunique()

# Concatanating dfs together step by step to avoid the memory overload of loading and concatanating all 7 at a time
dfs = [df_oct, df_nov]
full_data = pd.concat(dfs)

# Deleting unneeded dfs and clearing memory using the gc library
del df_oct
del df_nov
gc.collect()

# Reading csv file from S3 - Selecting specific columns
df_dec = pd.read_csv('s3://predictive-maintenance-bucket/data/2019-Dec.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code', 'category_id', 'brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_dec = df_dec.dropna(subset=['category_code', 'brand'])
df_dec = df_dec.loc[df_dec['event_type'] == 'purchase']
df_dec.nunique()

# Concatanating dfs together step by step to avoid the memory overload of loading and concatanating all 7 at a time
dfs = [full_data, df_dec]
full_data = pd.concat(dfs)

# Deleting unneeded dfs and clearing memory using the gc library
del df_dec
gc.collect()

# Reading csv file from S3 - Selecting specific columns
df_jan = pd.read_csv('s3://predictive-maintenance-bucket/data/2020-Jan.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code', 'category_id','brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_jan = df_jan.dropna(subset=['category_code', 'brand'])
df_jan = df_jan.loc[df_jan['event_type'] == 'purchase']
df_jan.nunique()

# Concatanating dfs together step by step to avoid the memory overload of loading and concatanating all 7 at a time
dfs = [full_data, df_jan]
full_data = pd.concat(dfs)

# Deleting unneeded dfs and clearing memory using the gc library
del df_jan
gc.collect()

# Reading csv file from S3 - Selecting specific columns
df_feb = pd.read_csv('s3://predictive-maintenance-bucket/data/2020-Feb.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code', 'category_id', 'brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_feb = df_feb.dropna(subset=['category_code', 'brand'])
df_feb = df_feb.loc[df_feb['event_type'] == 'purchase']
df_feb.nunique()

# Concatanating dfs together step by step to avoid the memory overload of loading and concatanating all 7 at a time
dfs = [full_data, df_feb]
full_data = pd.concat(dfs)

# Deleting unneeded dfs and clearing memory using the gc library
del df_feb
gc.collect()

# Reading csv file from S3 - Selecting specific columns
df_march = pd.read_csv('s3://predictive-maintenance-bucket/data/2020-Mar.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code', 'category_id', 'brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_march = df_march.dropna(subset=['category_code', 'brand'])
df_march = df_march.loc[df_march['event_type'] == 'purchase']
df_march.nunique()

# Concatanating dfs together step by step to avoid the memory overload of loading and concatanating all 7 at a time
dfs = [full_data, df_march]
full_data = pd.concat(dfs)

# Deleting unneeded dfs and clearing memory using the gc library
del df_march
gc.collect()

# Reading csv file from S3 - Selecting specific columns
df_apr = pd.read_csv('s3://predictive-maintenance-bucket/data/2020-Apr.csv', usecols=['user_id', 'user_session', 'event_time', 'category_code', 'category_id', 'brand', 'product_id', 'event_type'])
# Preliminary Cleaning
# Only keeping records that contain the category_code and brand in order to make more meaningful recommendations
df_apr = df_apr.dropna(subset=['category_code', 'brand'])
df_apr = df_apr.loc[df_apr['event_type'] == 'purchase']
df_apr.nunique()

# Concatanating dfs together step by step to avoid the memory overload of loading and concatanating all 7 at a time
dfs = [full_data, df_apr]
full_data = pd.concat(dfs)

# Deleting unneeded dfs and clearing memory using the gc library
del df_apr
gc.collect()

full_data = full_data[['user_id','user_session','event_time','category_code', 'category_id', 'brand', 'product_id']]
full_data['category'] = full_data['category_id'].astype(str) + '_' + full_data['category_code']# Creating category column
full_data.head()

full_data.nunique()

# Saving results in S3
full_data.to_csv('s3://myaws-capstone-bucket/eCommerce_purchase_data.csv', index=False)
