==> EDA/7_RFM_Segmentation.ipynb <==
Creating a RFM segmentations for the customers that purchased (during the 7 months of data) which were identified on the notebook:Customer_Behavior_Profile_Analysis.ipynb

RFM Definitions
- Recency (R): Days since last purchase
- Frequency (F): Total purchases the customer has
- Monetary Value (M): Total customer spent
    
- Using the following article as reference: https://towardsdatascience.com/find-your-best-customers-with-customer-segmentation-in-python-61d602f9eee6

# Loading basic needed libraries
import pandas as pd
import numpy as np
import gc
from functools import reduce
import datetime as dt
from datetime import date

# Loading libraries for S3 bucket connection
import boto3
import io
from io import StringIO,BytesIO, TextIOWrapper
import gzip

client = boto3.client('s3') 
resource = boto3.resource('s3') 

# Reading customers who purchased
main_custs = pd.read_csv('s3://myaws-capstone-bucket/data/customers_of_focus.csv')
main_custs.nunique()

# Verifying distribution of numeric columns
pd.options.display.float_format = '{:.2f}'.format
main_custs.describe()

# Creating RFM DF
rfm_df = main_custs[['user_id','days_since_last_purchase','total_purchases','total_spent']]# Keeping needed columns for RFM
rfm_df.columns = ['user_id','R','F','M']# renaming the columns
rfm_df.head()

rfm_df.describe()

# Creating RFM Scores based on Quantile distribution
Q = rfm_df.quantile(q=[0.25,0.5,0.75])
Q = Q.to_dict()

# Recency Score will be calculated inversely from Frequency and Monetary Values
# The lower the recency the higher the score should be 
def R_Score(x,p,d):
    if x >= d[p][0.25]:
        return 1
    elif x >= d[p][0.50]:
        return 2
    elif x >= d[p][0.75]: 
        return 3
    else:
        return 4
    
# The higher the Frequency and Monetary Values the higher the score should be
def FM_Score(x,p,d):
    if x >= d[p][0.25]:
        return 4
    elif x >= d[p][0.50]:
        return 3
    elif x >= d[p][0.75]: 
        return 2
    else:
        return 1
    

# Creating RFM quantile columns
rfm_df['R_quantile_score'] = rfm_df['R'].apply(R_Score, args=('R',Q,))
rfm_df['F_quantile_score'] = rfm_df['F'].apply(FM_Score, args=('F',Q,))
rfm_df['M_quantile_score'] = rfm_df['M'].apply(FM_Score, args=('M',Q,))

# RFM Score column creation
rfm_df['RFM_Score'] = rfm_df['R_quantile_score'] + rfm_df['F_quantile_score'] + rfm_df['M_quantile_score']

rfm_df.head()

rfm_df['RFM_Score'].describe()# Checking the RFM_Score distribution

rfm_df.head()

Creating Customer Segment Values:
- Top Customer - The customers with max RFM values
- High Value Customer - The customers that have RFM values larger than 50th percentile but less than the max score
- Mid Value Customer - The customers that have RFM values larger than 25th percentile but less than or equal to 50th percentile
- Low Value Customer - The customers that have RFM values less than 25th percentile 

# Creating customer segment based on RFM values
rfm_df.loc[(rfm_df['RFM_Score'] == 12),'user_segment'] = 'Top_Customer'
rfm_df.loc[(rfm_df['RFM_Score'] == 11),'user_segment'] = 'High_Value_Customer'
rfm_df.loc[(rfm_df['RFM_Score'] <= rfm_df['RFM_Score'].quantile(.75)),'user_segment'] = 'High_Value_Customer'
rfm_df.loc[(rfm_df['RFM_Score'] <= rfm_df['RFM_Score'].quantile(.50)),'user_segment'] = 'Mid_Value_Customer'
rfm_df.loc[(rfm_df['RFM_Score'] <= rfm_df['RFM_Score'].quantile(.25)),'user_segment'] = 'Low_Value_Customer'
rfm_df.head()

# Counting amount of users in each segment
rfm_df['user_segment'].value_counts()

# Saving Results in S3
rfm_df = rfm_df[['user_id','RFM_Score','user_segment']]
rfm_df.to_csv('s3://myaws-capstone-bucket/data/rfm_segment.csv',index=False)
